+++ R/simulation-spades.R
@@ -839,32 +839,23 @@ setGeneric(
 setMethod(
   "spades",
   signature(sim = "simList", cache = "missing"),
-  definition = function(sim,
-                        debug,
-                        progress,
-                        cache,
-                        .plotInitialTime,
-                        .saveInitialTime,
-                        notOlderThan,
-                        events,
-                        .plots,
-                        ...) {
-
-    # set the options; then set them back on exit
+  definition = function(sim, debug, progress, cache, .plotInitialTime, .saveInitialTime,
+                        notOlderThan, events, .plots, ...) {
+    ## set the options; then set them back on exit
     optsFromDots <- dealWithOptions(sim = sim)
     if (!is.null(optsFromDots$optsPrev)) {
-      # remove from `sim` as these should not be there
+      ## remove from `sim` as these should not be there
       rm(list = unique(names(optsFromDots$optionsAsProvided)), envir = envir(sim))
       on.exit({
-        # reset options in session
+        ## reset options in session
         options(optsFromDots$optsPrev)
-        # put them back in simList for reassessment during spades
+        ## put them back in simList for reassessment during spades
         if (exists("sim", inherits = FALSE))
           list2env(optsFromDots$optionsAsProvided, envir = envir(sim))
       }, add = TRUE)
     }
 
-    # loggingMessage helpers
+    ## loggingMessage helpers
     ._simNesting <- simNestingSetup(...)
     sim[["._simNesting"]] <- ._simNesting
 
@@ -878,9 +869,10 @@ setMethod(
       modNam <- getOption("spades.covr")
       tf <- tempfile();
       on.exit(unlink(tf))
-      cat(file = tf, paste('spades(sim, events = ',capture.output(dput(events)),', .plotInitialTime = ', .plotInitialTime, ')', collapse = "\n"))
+      cat(file = tf, paste('spades(sim, events = ', capture.output(dput(events)),
+                           ', .plotInitialTime = ', .plotInitialTime, ')', collapse = "\n"))
       sim$.mods[[modNam]]$sim <- sim
-      opts <- options("spades.covr2" = FALSE) # turn off this chunk 2nd time through
+      opts <- options("spades.covr2" = FALSE) ## turn off this chunk 2nd time through
       on.exit(options(opts), add = TRUE)
       aa <- covr::environment_coverage(sim$.mods[[modNam]], test_files = tf)
       rm(list = "sim", envir = sim$.mods[[modNam]])
@@ -902,11 +894,10 @@ setMethod(
           all(!grepl("writeToConsole", names(logging::getLogger()[["handlers"]])))
       } else {
         debug <- unlist(debug)
-
       }
     }
 
-    # need to recheck package loading because `simInit` may have been cached
+    ## need to recheck package loading because `simInit` may have been cached
     if (getOption("spades.loadReqdPkgs", TRUE)) {
       pkgs <- packages(sim)
       loadPkgs(pkgs)
@@ -964,7 +955,7 @@ setMethod(
         }
       }
 
-      # This sets up checking for memory leaks
+      ## This sets up checking for memory leaks
       if (is.null(sim@.xData[["._knownObjects"]])) {
         moduleNames <- unname(modules(sim))
         names(moduleNames) <- moduleNames
@@ -980,8 +971,8 @@ setMethod(
       .pkgEnv[["skipNamespacing"]] <- !getOption("spades.switchPkgNamespaces")
       .pkgEnv[["spades.keepCompleted"]] <- getOption("spades.keepCompleted", TRUE)
 
-      # Memory Use
-      # memory estimation of each event/sim
+      ## Memory Use
+      ## memory estimation of each event/sim
       if (getOption("spades.memoryUseInterval", 0) > 0) {
         if (requireNamespace("future", quietly = TRUE) &&
             requireNamespace("future.callr", quietly = TRUE)) {
@@ -995,7 +986,7 @@ setMethod(
         }
       }
 
-      # timeunits gets accessed every event -- this should only be needed once per simList
+      ## timeunits gets accessed every event -- this should only be needed once per simList
       sim@.xData$.timeunits <- timeunits(sim)
       on.exit({
         if (!.pkgEnv[["skipNamespacing"]])
@@ -1016,13 +1007,13 @@ setMethod(
           svdSimEnv$.sim <- sim # no copy of objects -- essentially 2 pointers throughout
           .pkgEnv$.cleanEnd <- NULL
         }
-        # For restarting R -- a few extra pieces, including saving the simList as the last thing
+        ## For restarting R -- a few extra pieces, including saving the simList as the last thing
         if (!is.null(sim$._restartRList)) {
           sim@simtimes[["current"]] <- sim@events[[1]]$eventTime
           sim$._restartRList$.spadesCall <- match.call()
 
           restartFormals <- formals(restartR)
-          # can change end(sim) back to original now because we are already ending
+          ## can change end(sim) back to original now because we are already ending
           end(sim) <- sim$._restartRList$endOrig
           restartR(
             sim = sim,
@@ -1101,10 +1092,10 @@ setMethod(
 
       sim@.xData[["._firstEventClockTime"]] <- Sys.time()
 
-      # This was introduced when sim@completed became an environment for speed purposes
-      # (list got slow as size increased)
-      # This is an attempt to deal with the expected behaviour of a list --
-      # i.e., delete it if this appears to be the original sim object again passed in
+      ## This was introduced when sim@completed became an environment for speed purposes
+      ## (list got slow as size increased)
+      ## This is an attempt to deal with the expected behaviour of a list --
+      ## i.e., delete it if this appears to be the original sim object again passed in
       if (length(sim@completed)) {
         existingCompleted <- sort(as.integer(ls(sim@completed, sorted = FALSE)))
         prevStart <- get(as.character(existingCompleted[1]), envir = sim@completed)
@@ -1119,7 +1110,7 @@ setMethod(
       }
 
       if (recoverMode > 0) {
-        rmo <- NULL # The recovery mode object
+        rmo <- NULL ## the recovery mode object
         allObjNames <- outputObjectNames(sim)
         if (is.null(allObjNames)) recoverMode <- 0
       }
@@ -1136,8 +1127,8 @@ setMethod(
         sim$.simFuture <- list()
       }
 
-      # There are some edge cases where there is an event scheduled before current time,
-      #   even though current time is after end time
+      ## There are some edge cases where there is an event scheduled before current time,
+      ##   even though current time is after end time
       if (length(sim@events)) {
         specialStart <- sim@events[[1]][["eventTime"]] < sim@simtimes[["current"]] &&
           sim@simtimes[["current"]] > sim@simtimes[["end"]]
@@ -1160,7 +1151,7 @@ setMethod(
         if (recoverMode > 0) {
           rmo <- recoverModePost(sim, rmo, recoverMode)
         }
-        # Conditional Scheduling -- adds only 900 nanoseconds per event, if none exist
+        ## Conditional Scheduling -- adds only 900 nanoseconds per event, if none exist
         if (exists("._conditionalEvents", envir = sim, inherits = FALSE)) {
           condEventsToOmit <- integer()
           for (condNum in seq(sim$._conditionalEvents)) {
@@ -1202,33 +1193,35 @@ setMethod(
       }
       sim@simtimes[["current"]] <- sim@simtimes[["end"]]
 
-      # For determining if clean ending to spades call
+      ## for determining if clean ending to spades call
       .pkgEnv$.cleanEnd <- TRUE
       return(invisible(sim))
     },
     warning = function(w) {
       w$message <- gsub("^In modCall\\(sim = sim.+\"]]\\): ", "", w$message)
-      if (grepl("NAs introduced by coercion", w$message))
-      if (newDebugging && requireNamespace("logging", quietly = TRUE)) {
-        logging::logwarn(paste0(collapse = " ", c(names(w), w)))
+      if (any(grepl("NAs introduced by coercion", w$message))) {
+        if (newDebugging && requireNamespace("logging", quietly = TRUE)) {
+          logging::logwarn(paste0(collapse = " ", c(names(w), w)))
+        }
       }
-      if (grepl("In .+:", w$message)) {
+      if (any(grepl("In .+:", w$message))) {
         warningSplitOnColon(w)
         invokeRestart("muffleWarning")
       } else {
         warning(w)
         tryCatch(invokeRestart("muffleWarning"), error = function(e) NULL)
       }
-
     },
     error = function(e) {
       if (newDebugging && requireNamespace("logging", quietly = TRUE)) {
-        if (debug > 0)
+        if (debug > 0) {
           logging::logerror(e)
+        }
       } else {
         fn <- get0("onError")
-        if (!is.null(fn))
+        if (!is.null(fn)) {
           fn(sim)
+        }
         stop(e)
       }
     },